requirements.txt
python-escpos[all]==3.1
python-telegram-bot==20.7
pyyaml==6.0.1
pillow==10.1.0
pydantic==2.5.3